requirements.txt
requests>=2.31.0
websockets
coinbase-advanced-py
python-dotenv
urllib3
zstandard
msal
azure-storage-queue
azure-storage-blob
azure-servicebus
azure-identity
pyarrow
